+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.59",
-  genVersion: "2.493.34",
-  userAgent: "speakeasy-sdk/typescript 0.7.59 2.493.34 1.0.0 ryan-lending",
+  sdkVersion: "0.7.60",
+  genVersion: "2.495.0",
+  userAgent: "speakeasy-sdk/typescript 0.7.60 2.495.0 1.0.0 ryan-lending",
 } as const;